Removed creating Variable from Symbol.
A Variable makes sense as part of a Grammar. Allowing a Symbol would
permit the following:

  pat = Pegarus.pattern(1) + :S

However, there's no way to link that Variable created with :S back
to the Variable used in the Grammar without reaching into that pattern.
It's just as simple and better to do the following

  g = Pegarus.grammar
  g.S = ...
  g.A = Pegarus.pattern(1) + g.S